Add: GlobalExceptionHandler

+++ src/main/java/com/marcos90s/goldSellerAPI/services/UsersService.java
@@ -3,6 +3,9 @@ package com.marcos90s.goldSellerAPI.services;
 import com.marcos90s.goldSellerAPI.dto.UsersRequestDTO;
 import com.marcos90s.goldSellerAPI.dto.UsersResponseDTO;
 import com.marcos90s.goldSellerAPI.entities.Users;
+import com.marcos90s.goldSellerAPI.exception.BadRequestException;
+import com.marcos90s.goldSellerAPI.exception.InternalServerErrorException;
+import com.marcos90s.goldSellerAPI.exception.NotFoundException;
 import com.marcos90s.goldSellerAPI.repository.UsersRepository;
 import com.marcos90s.goldSellerAPI.utils.Utils;
 import jakarta.transaction.Transactional;
@@ -18,13 +21,20 @@ public class UsersService {
     UsersRepository usersRepository;
 
     public UsersResponseDTO createUser(UsersRequestDTO dto) {
+        if (usersRepository.existsByEmail(dto.getEmail())){
+            throw new BadRequestException("Email is already exists!");
+        }
         Users user = new Users();
         user.setName(dto.getName());
         user.setEmail(dto.getEmail());
         user.setPassword(dto.getPassword());
         user.setRole(dto.getRole());
 
-        usersRepository.save(user);
+        try {
+            usersRepository.save(user);
+        } catch (Exception e) {
+            throw new InternalServerErrorException("Error while saving user!");
+        }
         return mapToResponseDTO(user);
     }
 
@@ -37,20 +47,34 @@ public class UsersService {
     @Transactional
     public UsersResponseDTO getUserById(String id) {
         Users user = usersRepository.findByIdWithTransactions(id)
-                .orElseThrow(() -> new RuntimeException("Usuário não encontrado"));
+                .orElseThrow(() -> new NotFoundException("User not Found!"));
 
         return mapToResponseDTO(user);
     }
 
     public void deleteById(String id) {
-        usersRepository.deleteById(id);
+        if (!usersRepository.existsById(id)){
+            throw new NotFoundException("User not Found!");
+        }
+        try {
+            usersRepository.deleteById(id);
+        } catch (Exception e) {
+            throw new InternalServerErrorException("Error while deleting user!");
+        }
     }
 
     public UsersResponseDTO updateUser(String id, UsersRequestDTO dto) {
+        if(!usersRepository.existsById(id)){
+            throw new NotFoundException("User not Found!");
+        }
         Users entity = usersRepository.getReferenceById(id);
         if (dto.getName() != null) entity.setName(dto.getName());
         if (dto.getPassword() != null) entity.setPassword(dto.getPassword());
-        usersRepository.save(entity);
+        try {
+            usersRepository.save(entity);
+        } catch (Exception e) {
+            throw new InternalServerErrorException("Error while saving user!");
+        }
         return mapToResponseDTO(entity);
     }
